Seperated Employee and User controller
