+++ src/main/java/com/ftms/ftmsapi/config/SecurityConfig.java
@@ -79,7 +79,6 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
                         "/**/*.css",
                         "/**/*.js")
                 .permitAll()
-
                 .antMatchers("/api/auth/**", "/api/recovery/**")
                     .permitAll()
                 .anyRequest()
